+++ sg-mini-firmware/src/DeviceSetting/SettingManager/WaterSettingManager/ModeSetting/ScheduleModeSetting/ScheduleModeSetting.cpp
@@ -3,7 +3,7 @@
 
 ScheduleModeSetting::ScheduleModeSetting() {}
 
-ScheduleModeSetting::ScheduleModeSetting(unsigned long scheduleDuration, float targetHumidity) {
+ScheduleModeSetting::ScheduleModeSetting(unsigned long scheduleDuration, uint8_t targetHumidity) {
     setScheduleDuration(scheduleDuration);
     setTargetHumidity(targetHumidity);
 }
@@ -12,8 +12,8 @@ void ScheduleModeSetting::setScheduleDuration(unsigned long scheduleDuration) {
     this->scheduleDuration = scheduleDuration;
 }
 
-void ScheduleModeSetting::setTargetHumidity(float targetHumidity) {
-    if (Helper::isInRange(targetHumidity, 0.0, 100.0))
+void ScheduleModeSetting::setTargetHumidity(uint8_t targetHumidity) {
+    if (Helper::isInRange(targetHumidity, 0, 100))
         this->targetHumidity = targetHumidity;
 }
 
@@ -21,7 +21,7 @@ unsigned long ScheduleModeSetting::getScheduleDuration() {
     return this->scheduleDuration;
 }
 
-float ScheduleModeSetting::getTargetHumidity() {
+uint8_t ScheduleModeSetting::getTargetHumidity() {
     return this->targetHumidity;
 }
